+++ lib/src/solid/chk_exists_and_has_acl.dart
@@ -36,7 +36,6 @@ import 'package:solidpod/src/solid/api/common_permission.dart';
 import 'package:solidpod/src/solid/api/rest_api.dart';
 import 'package:solidpod/src/solid/constants/common.dart';
 import 'package:solidpod/src/solid/solid_func_call_status.dart';
-import 'package:solidpod/src/solid/utils/exceptions.dart';
 import 'package:solidpod/src/solid/utils/misc.dart';
 
 /// Check [fileName] exists and has the associated ACL file. Requires user
@@ -63,11 +62,8 @@ Future<SolidFunctionCallStatus> chkExistsAndHasAcl({
 }) async {
   if (!context.mounted) return SolidFunctionCallStatus.contextNotMounted;
 
-  if (!await checkLoggedIn()) {
-    throw NotLoggedInException(
-      'User must be logged in to check resource ACL. '
-      'Please authenticate before calling chkExistsAndHasAcl().',
-    );
+  if (!await isUserLoggedIn()) {
+    return SolidFunctionCallStatus.notLoggedIn;
   }
 
   if (!context.mounted) return SolidFunctionCallStatus.contextNotMounted;