+++ lib/src/solid/revoke_permission.dart
@@ -188,7 +188,7 @@ Future<SolidFunctionCallStatus> revokePermission({
       // Owner
       // [20251029 jesscmoore] Assumes user = owner and uses
       // AuthDataManager.getWebId() to fetch user webId
-      final ownerLogFileUrl = await getFileUrl(logFilePath, ownerWebId);
+      final ownerLogFileUrl = await getFileUrl(logFilePath, webId: ownerWebId);
 
       // Granter
       // [20251029 jesscmoore] Assumes user = granter and uses
@@ -215,7 +215,7 @@ Future<SolidFunctionCallStatus> revokePermission({
           .contains(recipientType)) {
         if (await checkPodInitialised(removerWebId)) {
           final receiverLogFileUrl =
-              await getFileUrl(logFilePath, removerWebId);
+              await getFileUrl(logFilePath, webId: removerWebId);
           await addPermLogLine(
             logFileUrl: receiverLogFileUrl,
             logEntry: logEntryRes,