+++ lib/src/solid/utils/misc.dart
@@ -43,14 +43,16 @@ import 'package:path/path.dart' as path;
 import 'package:rdflib/rdflib.dart';
 import 'package:solid_auth/solid_auth.dart' show genDpopToken, logout;
 
-import 'package:solidpod/solidpod.dart';
 import 'package:solidpod/src/solid/api/rest_api.dart';
 import 'package:solidpod/src/solid/constants/common.dart';
+import 'package:solidpod/src/solid/constants/path_type.dart';
 import 'package:solidpod/src/solid/constants/schema.dart';
 import 'package:solidpod/src/solid/constants/web_acl.dart';
 import 'package:solidpod/src/solid/revoke_permission_to_recipients.dart';
-import 'package:solidpod/src/solid/utils/authdata_manager.dart'
-    show AuthDataManager;
+import 'package:solidpod/src/solid/utils/app_info.dart';
+import 'package:solidpod/src/solid/utils/authdata_manager.dart';
+import 'package:solidpod/src/solid/utils/exceptions.dart';
+import 'package:solidpod/src/solid/utils/key_manager.dart';
 import 'package:solidpod/src/solid/utils/permission.dart';
 import 'package:solidpod/src/solid/utils/rdf.dart';
 
@@ -97,19 +99,19 @@ String decryptData(
     Encrypter(AES(key, mode: mode)).decrypt(Encrypted.from64(encData), iv: iv);
 
 /// Load and parse a private TTL file from POD
-Future<Map<String, dynamic>> loadPrvTTL(String fileUrl) async {
-  // final fileUrl = await getFileUrl(filePath);
-  try {
-    if (await checkResourceStatus(fileUrl) == ResourceStatus.exist) {
-      final rawContent = await fetchPrvFile(fileUrl);
-      return parseTTL(rawContent);
-    } else {
-      return {};
-    }
-  } on Exception catch (e) {
-    throw Exception(e);
-  }
-}
+// Future<Map<String, dynamic>> loadPrvTTL(String fileUrl) async {
+//   // final fileUrl = await getFileUrl(filePath);
+//   try {
+//     if (await checkResourceStatus(fileUrl) == ResourceStatus.exist) {
+//       final rawContent = await fetchPrvFile(fileUrl);
+//       return parseTTL(rawContent);
+//     } else {
+//       return {};
+//     }
+//   } on Exception catch (e) {
+//     throw Exception(e);
+//   }
+// }
 
 /// Read the encryption key file content for display purposes.
 ///
@@ -117,20 +119,24 @@ Future<Map<String, dynamic>> loadPrvTTL(String fileUrl) async {
 /// making it suitable for accessing files outside the appname/data directory.
 ///
 /// Returns the raw TTL content of the encryption key file.
-Future<String> readEncryptionKeyContent() async {
-  final encKeyPath = await getEncKeyPath();
-  final encKeyUrl = await getFileUrl(encKeyPath);
+// Future<String> readEncryptionKeyContent() async {
+//   final encKeyPath = await getEncKeyPath();
+//   final encKeyUrl = await getFileUrl(encKeyPath);
 
-  try {
-    if (await checkResourceStatus(encKeyUrl) == ResourceStatus.exist) {
-      return await fetchPrvFile(encKeyUrl);
-    } else {
-      throw Exception('Encryption key file does not exist at: $encKeyPath');
-    }
-  } on Exception catch (e) {
-    throw Exception('Failed to read encryption key file: $e');
-  }
-}
+//   try {
+//     if (await checkResourceStatus(encKeyUrl) == ResourceStatus.exist) {
+//       return utf8.decode(
+//         await getResource(encKeyUrl),
+//       );
+
+//       return await fetchPrvFile(encKeyUrl);
+//     } else {
+//       throw Exception('Encryption key file does not exist at: $encKeyPath');
+//     }
+//   } on Exception catch (e) {
+//     throw Exception('Failed to read encryption key file: $e');
+//   }
+// }
 
 /// Generates a public key block from a given key content.
 String genPubKeyStr(String pubKeyContent) =>
@@ -155,22 +161,25 @@ String getUniqueIdResUrl(String resourceUrl, String receiverWebId) {
 }
 
 /// From a given resource path [resourcePath] create its URL
-/// [isContainer] should be true if the resource is a directory, otherwise false
+/// [resourcePath] should be relative to the POD, e.g., `myapp/data/abc.ttl`.
+/// [isFile] should be true if the resource is a file, otherwise false if it
+/// is a directory.
+/// [webId] (optionally) specifies the resource's POD (could be external).
 /// returns the full resource URL
 
 Future<String> _getResourceUrl(
-  String resourcePath,
-  bool isContainer, [
-  String? extWebId,
-]) async {
-  // Check if resource url is needed for an external webId
-  final webId = extWebId ?? await AuthDataManager.getWebId();
-  assert(webId != null);
-  assert(webId!.contains(profCard));
+  String resourcePath, {
+  required bool isFile,
+  String? webId,
+}) async {
+  final wId = webId ?? await AuthDataManager.getWebId();
 
-  final resourceUrl = webId!.replaceAll(profCard, resourcePath);
+  assert(wId != null);
+  assert(wId!.contains(profCard));
 
-  if (isContainer && !resourceUrl.endsWith('/')) {
+  final resourceUrl = wId!.replaceAll(profCard, resourcePath);
+
+  if (!isFile && !resourceUrl.endsWith('/')) {
     return '$resourceUrl/';
   }
 
@@ -178,12 +187,12 @@ Future<String> _getResourceUrl(
 }
 
 /// Create the URL for a file
-Future<String> getFileUrl(String filePath, [String? extWebId]) async =>
-    await _getResourceUrl(filePath, false, extWebId);
+Future<String> getFileUrl(String filePath, {String? webId}) async =>
+    await _getResourceUrl(filePath, isFile: true, webId: webId);
 
 /// Create the URL for a directory (container)
-Future<String> getDirUrl(String dirPath, [String? extWebId]) async =>
-    await _getResourceUrl(dirPath, true, extWebId);
+Future<String> getDirUrl(String dirPath, {String? webId}) async =>
+    await _getResourceUrl(dirPath, isFile: false, webId: webId);
 
 /// Get resource Url from a filename, with different options for how
 /// the filename is provided. If [isExternalRes] or [isFileUrl] is
@@ -264,7 +273,7 @@ Future<String> getEncTTLStr(
   String? inheritKeyFrom,
 }) async {
   final triples = {
-    URIRef(await getFileUrl(filePath, extWebId)): {
+    URIRef(await getFileUrl(filePath, webId: extWebId)): {
       solidTermsNS.ns.withAttr(pathPred): filePath,
       solidTermsNS.ns.withAttr(ivPred): iv.base64,
       if (inheritKeyFrom != null)
@@ -330,9 +339,8 @@ Future<String?> getWebId() async => AuthDataManager.getWebId();
 /// Check if the local storage has authentication
 /// details of the user and also check whether the
 /// access token is expired or not
-/// returns boolean
 
-Future<bool> checkLoggedIn() async {
+Future<bool> isUserLoggedIn() async {
   final webId = await AuthDataManager.getWebId();
 
   if (webId != null && webId.isNotEmpty) {
@@ -546,8 +554,7 @@ Future<void> initPod(
 }) async {
   // Check if the user has logged in
 
-  final loggedIn = await checkLoggedIn();
-  if (!loggedIn) {
+  if (!await isUserLoggedIn()) {
     throw NotLoggedInException('Can not initialise POD without logging in');
   }
 
@@ -787,4 +794,62 @@ bool isDir(String path) {
   } else {
     return false;
   }
+}
+
+/// Generate the URL of resource according to its path and the type of the path.
+
+Future<String> generateResourceUrlFromPath({
+  required String resourcePath,
+  required PathType pathType,
+  bool isFile = true,
+  String? webId,
+}) async {
+  final func = isFile ? getFileUrl : getDirUrl;
+  switch (pathType) {
+    case PathType.absoluteUrl:
+      return resourcePath;
+
+    case PathType.relativeToPod:
+      return await func(
+        resourcePath,
+        webId: webId,
+      );
+
+    case PathType.relativeToApp:
+      return await func(
+        [appDirName, resourcePath].join('/'),
+        webId: webId,
+      );
+
+    case PathType.relativeToData:
+      return await func(
+        [await getDataDirPath(), resourcePath].join('/'),
+        webId: webId,
+      );
+  }
+}
+
+/// Extract resource path from its URL
+/// path format:
+/// - appDir/path/to/file
+/// - appDir/path/to/dir/
+
+Future<String> extractResourcePathFromUrl(
+  String resourceUrl, {
+  bool isFile = true,
+}) async {
+  // See https://api.dart.dev/dart-core/Uri-class.html for details
+
+  final segments = Uri.parse(resourceUrl).pathSegments;
+
+  final path = segments.getRange(1, segments.length).join('/');
+
+  return !(isFile || path.endsWith('/')) ? '$path/' : path;
+}
+
+/// Generate the Web ID of from resource URL
+
+Future<String> generateWebIdFromResourceUrl(String resourceUrl) async {
+  final uri = Uri.parse(resourceUrl);
+  return [uri.origin, uri.pathSegments.first, profCard].join('/');
 }