+++ lib/src/solid/utils/permission.dart
@@ -27,6 +27,7 @@
 library;
 
 import 'package:flutter/foundation.dart' show debugPrint;
+import 'dart:convert';
 
 import 'package:rdflib/rdflib.dart' show URIRef, Namespace;
 
@@ -170,9 +171,11 @@ Future<Map<dynamic, dynamic>> readAcl(
 ]) async {
   final resourceAclUrl = getResAclFile(resourceUrl, isFile);
 
-  final aclContent = await fetchPrvFile(resourceAclUrl);
-
+  final aclContent = utf8.decode(
+    await getResource(resourceAclUrl),
+  );
   debugPrint('readAcl(): $aclContent');
+
   return parseACL(aclContent);
 }
 
@@ -180,7 +183,9 @@ Future<Map<dynamic, dynamic>> readAcl(
 ///
 /// Returns a Future that completes with a List containing the list of WebIDs.
 Future<List<dynamic>> readGroupTtl(String groupFileUrl) async {
-  final groupContent = await fetchPrvFile(groupFileUrl);
+  final groupContent = utf8.decode(
+    await getResource(groupFileUrl),
+  );
 
   final groupDataMap = parseTTLMap(groupContent);